requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
httpx==0.27.0
pydantic==2.6.4
python-dateutil==2.9.0.post0
starlette==0.36.3
tortoise-orm==0.20.0
